+++ src/UniversalPresentationFramework.Renderers.SkiaSharp.Vulkan/SkiaRendererVulkanProvider.cs
@@ -245,7 +245,7 @@ namespace Wodsoft.UI.Renderers
             return x.Item1 > y.Item1 ? -1 : 1;
         }
 
-        public override IRenderBitmapContext CreateRenderBitmapContext(int pixelWidth, int pixelHeight, double dpiX, double dpiY, PixelFormat pixelFormat)
+        public override IRenderBitmapContext CreateRenderBitmapContext(int pixelWidth, int pixelHeight, float dpiX, float dpiY, PixelFormat pixelFormat)
         {
             SKColorSpace colorSpace;
             switch (pixelFormat.ColorSpace)
